api/pyproject.toml
[project]
name = "turf-api"
version = "0.1.0"
description = "Turf API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.23.2",
    "SQLAlchemy==2.0.23",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "geoalchemy2==0.14.4",
    "python-dotenv==1.0.1",
    "pydantic==1.10.14",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.0.1",
    "python-multipart==0.0.9",
    "httpx>=0.24.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
